Misc edits
Setting initial values for optimization variables.
Removing tests that fail for known reasons.

+++ pyomo/neos/tests/test_neos.py
@@ -19,6 +19,7 @@
 import os
 import json
 from os.path import abspath, dirname
+import os.path
 currdir = dirname(abspath(__file__))
 
 import pyutilib.th as unittest
@@ -139,8 +140,11 @@ class TestSolvers_script_min(unittest.TestCase):
     def test_knitro(self):
         self._run('knitro')
 
-    def test_lbfgsb(self):
-        self._run('l-bfgs-b')
+    # This solver does not return
+    # a solution when all variables are
+    # projected to be at their bounds.
+    #def test_lbfgsb(self):
+    #    self._run('l-bfgs-b')
 
     def test_lancelot(self):
         self._run('lancelot')
@@ -160,9 +164,13 @@ class TestSolvers_script_min(unittest.TestCase):
     def test_mosek(self):
         self._run('mosek')
 
-    def test_ooqp(self):
-        self._run('ooqp')
+    # This solver doesn't give the same answer for
+    # min f(x) and max -f(x)
+    #def test_ooqp(self):
+    #    self._run('ooqp')
 
+    # The simple tests aren't complementarity 
+    # problems
     #def test_path(self):
     #    self._run('path')
 
@@ -217,14 +225,18 @@ class TestSolvers_cmd_min(TestSolvers_script_min):
                 data['Solver'][0]['Status'], 'ok')
             self.assertAlmostEqual(
                 data['Solution'][1]['Status'], 'optimal')
+            if 'x' in data['Solution'][1]['Variable']:
+                self.assertAlmostEqual(
+                    data['Solution'][1]['Variable']['x']['Value'], 0.5)
+            else:
+                self.fail("Expected nonzero solution variables")
             if 'o' in data['Solution'][1]['Objective']:
                 self.assertAlmostEqual(
                     data['Solution'][1]['Objective']['o']['Value'], self.objective)
-            self.assertAlmostEqual(
-                data['Solution'][1]['Variable']['x']['Value'], 0.5)
         finally:
             cleanup()
-            os.remove(results)
+            if os.path.exists(results):
+                os.remove(results)
 
 
 @unittest.category('nightly')
